DESCW-1105 Restored Reports for Individual Project.
