--- backend/src/controllers/reports/Tab_16_rpt_P_QuarterlyReview_multiyrsample.js
@@ -1,52 +1,0 @@
-const useCommonComponents = require("../useCommonComponents/index");
-const useController = require("../useController/index");
-const model = require("@models/reports/Tab_16_rpt_P_QuarterlyReview_multiyrsample");
-const utils = require("./helpers");
-const what = { single: "report", plural: "reports" };
-const controller = useController(model, what);
-
-// Template and data reading
-const cdogs = useCommonComponents("cdogs");
-const { getReport, getDocumentApiBody, pdfConfig } = utils;
-controller.getReport = getReport;
-
-/**
- * Get a Project rollup Report for a specific array of portfolio.
- *
- * @param   {FastifyRequest} request FastifyRequest is an instance of the standard http or http2 request objects.
- * @param   {FastifyReply}   reply   FastifyReply is an instance of the standard http or http2 reply types.
- * @returns {object}
- */
-controller.Tab_16_rpt_P_QuarterlyReview_multiyrsample = async (request, reply) => {
-  controller.userRequires(request, "PMO-Reports-Capability", reply);
-  try {
-    // Get the data from the database.
-    const getDate = async () => new Date();
-
-    const result = {
-      report_date: await getDate(),
-    };
-
-    const body = await getDocumentApiBody(
-      result,
-      "Tab_16_rpt_P_QuarterlyReview_multiyrsample.docx"
-    );
-    const pdf = await cdogs.api.post("/template/render", body, pdfConfig);
-
-    // Inject the pdf data into the request object
-    request.data = pdf;
-
-    if (!result) {
-      reply.code(404);
-      return { message: `The ${what.single} with the specified id does not exist.` };
-    } else {
-      return result;
-    }
-  } catch (err) {
-    console.error(`ERROR: ${err}`);
-    reply.code(500);
-    return { message: `There was a problem looking up this Project rollup Report.` };
-  }
-};
-
-module.exports = controller;
