+++ backend/src/models/reports/Tab_18_rpt_P_StatusSummary.js
@@ -1,2 +1,17 @@
-// Because all of the queries are coming from the global project, we can do this.
-module.exports = require("@models/reports/project");
+const {
+  getProjectById,
+  projectStatusReport,
+  getMilestones,
+  getStrategicAlignment,
+  getProjectStatuses,
+  getLessonsLearned,
+} = require("@models/reports/useProject");
+
+module.exports = {
+  getProjectById,
+  projectStatusReport,
+  getMilestones,
+  getStrategicAlignment,
+  getProjectStatuses,
+  getLessonsLearned,
+};
